requirements.txt
pandas
geopandas
matplotlib
seaborn
mapclassify
folium
PyYAML
unidecode
tqdm
pyarrow
scikit-learn